+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to a section with the version name.
 ## Unreleased Changes
 
 * Validate address details on submitting `Standard Tax Invoice`.
+* Fix error on invoice creation showing validation message
 
 ## 0.50.3